refactor(raylib): PrimitiveRenderer 拆分 S4a——Decal 抽为 RaylibDecalProjectorRenderer（Refs #1032）
- 新建 RaylibDecalProjectorRenderer（internal）：贴花 shader/材质/loc/常量与 DrawTexturedDecal 逻辑整体迁入，自行管理生命周期
- 新建 RaylibMaterialDrawState（internal static）：ResolveBlendMode/TryBeginAuthorBlendMode 单一来源，decal 与 billboard 路径共用，宿主私有副本删除
- 宿主 facade：公开 API/诊断属性/计数语义/异常契约不变；TryDrawDecalItem 仅留 AssetKind 检查+计数+转发
- DefaultVegetationAlphaCutoff 升 internal（decal cutout cutoff 单一来源）
- 3598 → 3322 行

验收：构建 0 错误；测试 66/67（唯一失败为预存无关项）；decal_projection 场景贴花正常（last frame 3）；vegetation_cutout 与 S3 基线像素级一致（maxdiff=1）；lighting 差异为场景昼夜相位时间漂移（同代码 run-to-run 亦有 maxdiff=12）
